+++ mailpoet/assets/js/src/newsletters/store/default-state.js
@@ -1,13 +1,10 @@
-import { storeName } from './constants';
-
 
 export const defaultNewsletterState = {
-    newsletterListing: {
-        data: [],
-        meta: { count: 0, groups: [], filters: { segment: [] } },
-    },
+    
+    newsletterStandardRows: [],
     errors: [],
     isFetching: false,
+    currentNewsletterType: "standard",
     mailpoet_newsletters_count: window.mailpoet_newsletters_count,
     wordpressRoles: window.wordpress_editable_roles_list,
